:white_check_mark: getRecentEightDates 테스트 함수 추가
